feat: generic transfer receipt
