+++ sn_networking/Cargo.toml
@@ -32,6 +32,7 @@ rmp-serde = "1.1.1"
 serde = { version = "1.0.133", features = [ "derive", "rc" ]}
 sn_protocol = { path = "../sn_protocol", version = "0.6.9" }
 sn_dbc = { version = "20.0.0", features = ["serdes"] }
+sn_transfers = { path = "../sn_transfers", version = "0.11.11" }
 thiserror = "1.0.23"
 tokio = { version = "1.32.0", features = ["io-util", "macros", "parking_lot", "rt", "sync", "time"] }
 tracing = { version = "~0.1.26" }
