+++ sn_networking/src/error.rs
@@ -23,7 +23,6 @@ pub(super) type Result<T, E = Error> = std::result::Result<T, E>;
 #[derive(Debug, Error)]
 #[allow(missing_docs)]
 pub enum Error {
-    /// Could not acquire a Semaphore permit.
     #[error("Could not acquire a Semaphore permit.")]
     CouldNotAcquireSemaphorePermit(#[from] tokio::sync::AcquireError),
