+++ sn_transfers/src/wallet/local_store.rs
@@ -381,20 +381,6 @@ impl LocalWallet {
         transfer.utxos(self.key.secret_key())
     }
 
-    /// Takes in a list of transfers and tries to decipher them one by one until
-    /// we find one that is for us.
-    /// Return the UTXOs if we find one.
-    /// Return an error if we find none.
-    pub fn unwrap_transfer_for_us(&self, payment: Vec<Transfer>) -> Result<Vec<Utxo>> {
-        for tr in payment {
-            match self.unwrap_transfer(tr) {
-                Ok(utxos) => return Ok(utxos),
-                Err(_) => continue, // cannot decipher, not ours
-            }
-        }
-        Err(Error::NotRecipient)
-    }
-
     pub fn derive_key(&self, derivation_index: &DerivationIndex) -> DerivedKey {
         self.key.derive_key(derivation_index)
     }
